Rename Bolt|EmbeddedPoint to Bolt|EmbeddedValue.
Fits the notion from Bolts and Embedded itself way better (see support for date types).
